requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain>=0.0.335
langchain-groq>=0.3.2
langchain-community>=0.0.23
langchain-core>=0.3.49
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
unidecode>=1.3.7
werkzeug>=2.3.7
gunicorn>=21.2.0
Pillow>=10.0.0
PyMuPDF>=1.22.0
boto3>=1.34.0
botocore>=1.34.0